fix: DB 저장 안정성 개선 및 한국시간 기준 로그 롤링 추가
  DB 저장 문제 해결:
  - numpy float64 타입을 Python float으로 변환 (schema "np" 오류 수정)
  - 스레드 안전성 추가 (RLock으로 동시 접근 방지)
  - SSH 터널 keepalive 설정 (30초 간격)
  - DB 재시도 횟수 2회 → 3회 증가
  - 알림 저장 순서 변경 (DB 저장 성공 후 텔레그램 발송)
  - 상세 로깅 추가 (저장 성공/실패/스킵 상태 출력)

  로깅 시스템:
  - 한국 시간 자정 기준 일별 자동 로그 롤링
  - logs/bot.log → logs/bot.log.YYYYMMDD 형식
  - 30일치 자동 보관

  실행 방법 변경:
  - 이전: nohup python -m us_oneil_simple > logs/bot_$(date).log 2>&1 &
  - 변경: nohup python -m us_oneil_simple &

+++ src/us_oneil_simple/patterns/pivot.py
@@ -48,10 +48,10 @@ def detect_pivot_breakout(
                 'ticker': ticker,
                 'pattern': '피벗돌파',
                 'market': market,
-                'resistance': resistance,
-                'current_price': current_price,
-                'breakout_pct': round(breakout_pct, 2),
-                'volume_surge': round(volume_surge, 2)
+                'resistance': float(resistance),  # numpy → Python float
+                'current_price': float(current_price),  # numpy → Python float
+                'breakout_pct': float(round(breakout_pct, 2)),
+                'volume_surge': float(round(volume_surge, 2))
             }
 
             if market == 'KR' and stock_name:
